move and fix the rest
