+++ .eslintrc.js
@@ -7,25 +7,29 @@ module.exports = {
     'airbnb-base',
   ],
   globals: {
-    '$': "readonly",
+    $: 'readonly',
   },
   parserOptions: {
     ecmaVersion: 12,
     sourceType: 'module',
   },
   rules: {
-    "max-len": "off",
-    "prefer-destructuring": "off",
-    "no-alert": "off",
-    "no-console": "off",
-    "max-classes-per-file": "off",
-    "no-new": "off",
+    'max-len': 'off',
+    'prefer-destructuring': 'off',
+    'no-alert': 'off',
+    'no-console': 'off',
+    'max-classes-per-file': 'off',
+    'no-new': 'off',
   },
   settings: {
-   "import/resolver": {
-     "node": {
-       "paths": ["app/javascript"],
-     },
-   },
- },
+    'import/resolver': {
+      node: {
+        paths: ['app/javascript/'],
+        extensions: [
+          '.js',
+          '.js.erb',
+        ],
+      },
+    },
+  },
 };
