+++ app/javascript/src/appsignal.js.erb
