+++ app/javascript/src/direct_uploads.js
